+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.module.ts
@@ -6,11 +6,12 @@ import { AttributeDescriptorTooltipPipeModule } from "../../util/pipes/Attribute
 import { FilterMetricDataBySearchTermPipe } from "./filterMetricDataBySearchTerm.pipe"
 import { MetricChooserComponent } from "./metricChooser.component"
 import { NodeSelectionService } from "./nodeSelection.service"
+import { MetricChooserValueModule } from "./metricChooserValue/metricChooserValue.module"
 
 @NgModule({
     declarations: [MetricChooserComponent, FilterMetricDataBySearchTermPipe],
     exports: [MetricChooserComponent],
-    imports: [CommonModule, MaterialModule, FormsModule, AttributeDescriptorTooltipPipeModule],
+    imports: [CommonModule, MaterialModule, FormsModule, AttributeDescriptorTooltipPipeModule, MetricChooserValueModule],
     providers: [NodeSelectionService]
 })
 export class MetricChooserModule {}