+++ visualization/app/codeCharta/ui/metricChooser/metricChooserType/metricChooserType.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, ViewEncapsulation } from "@angular/core"
 import { Store } from "@ngrx/store"
 import { Observable, map } from "rxjs"
-import { AttributeTypes, PrimaryMetrics, CcState, Node, CodeMapNode } from "../../../codeCharta.model"
+import { PrimaryMetrics, CcState, Node, CodeMapNode, AttributeTypes } from "../../../codeCharta.model"
 import { createAttributeTypeSelector } from "./createAttributeTypeSelector.selector"
 import { NodeSelectionService } from "../nodeSelection.service"
 import { isLeaf } from "../../../util/codeMapHelper"
@@ -12,8 +12,8 @@ import { isLeaf } from "../../../util/codeMapHelper"
     encapsulation: ViewEncapsulation.None
 })
 export class MetricChooserTypeComponent implements OnInit {
-    @Input() metricFor: keyof PrimaryMetrics
-    @Input() attributeType: keyof AttributeTypes
+    @Input({ required: true }) metricFor: keyof PrimaryMetrics
+    @Input() attributeType: keyof AttributeTypes = "nodes"
 
     isNodeALeaf$: Observable<boolean>
     attributeType$: Observable<string>