+++ visualization/app/codeCharta/ui/metricChooser/metricChooserValue/metricChooserValue.module.ts
@@ -2,9 +2,10 @@ import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { MetricChooserTypeModule } from "../metricChooserType/metricChooserType.module"
 import { MetricChooserValueComponent } from "./metricChooserValue.component"
+import { RoundedBoxModule } from "../../ribbonBar/roundedBox/roundedBox.module"
 
 @NgModule({
-    imports: [CommonModule, MetricChooserTypeModule],
+    imports: [CommonModule, MetricChooserTypeModule, RoundedBoxModule],
     declarations: [MetricChooserValueComponent],
     exports: [MetricChooserValueComponent]
 })